feat: added check if git is not clean

+++ scripts/build_players.sh
@@ -1,5 +1,12 @@
 #!/bin/sh
 
+is_clean=$(git update-index --refresh)
+
+if [ ! -z "$is_clean" ]; then
+  echo "git workspace is not clean"
+  exit 1
+fi
+
 rm -rf ./temp
 mkdir ./temp
